@@ -7,25 +7,28 @@ export const actions: Actions = {
 		const nameEntry = data.get('name');
 		if (nameEntry === null) {
 			return fail(400, {
-				missing: 'name'
+				missing: 'name',
 			});
 		}
 
 		const tallyEntry = data.get('tally');
 		if (tallyEntry === null) {
 			return fail(400, {
-				missing: 'tally'
+				missing: 'tally',
 			});
 		}
+
 		const name = `${nameEntry}`;
 		const tally = +tallyEntry;
 
+		const person = {
+			name,
+			tally,
+		};
+
 		return {
 			success: true,
-			person: {
-				name,
-				tally
-			}
+			person,
 		};
-	}
+	},
 };
