@@ -1,3 +1,9 @@
 import type { Present } from './types';
 
-export default (presents: Present[]) => presents.reduce((sum, present) => sum + present.weight, 0);
+export default (presents: Iterable<Present>) => {
+	let weight = 0;
+	for (const present of presents) {
+		weight += present.weight;
+	}
+	return weight;
+};
